projects/airflow/pyproject.toml
[tool.poetry]
name = "data-pipelines"
version = "0.0.1"
description = "pipelines gathering data to support research at ROCS"
authors = ["Jakob J. Kolb <jakob.j.kolb@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.10"
returns = "^0.16.0"
apache-airflow = "^2.1.4"
fsspec = "^2021.7.0"
ramda = "^2021.10.24"
yoyo-migrations = "^7.3.2"
pyhumps = "^3.0.2"
openpyxl = "^3.0.9"
great-expectations = "^0.13.26"
polars = "0.9.12"
psycopg2-binary = "^2.9.1"
celery = "5.1.2"
py7zr = "^0.16.1"
database = {path = "../../libs/database", develop=true}

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = { version = "*", allow-prereleases = true } # according to https://github.com/peterjc/flake8-black#python-package-management
flake8-black = "^0.2.1"
pytest = "^6.2.4"
testfixtures = "^6.17.1"
pytest-xdist = "^2.3.0"
mypy = "^0.910"
types-requests = "^2.25.0"
pytest-dotenv = "^0.5.2"
types-Markdown = "^3.3.0"
types-PyMySQL = "^1.0.0"
types-croniter = "^0.1.4"
types-python-dateutil = "^0.1.4"
types-setuptools = "^57.0.0"
types-tabulate = "^0.1.1"
types-toml = "^0.1.3"
responses = "^0.14.0"

[tool.pytest]
env_files =".env"
looponfailroots="dags"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

